fix: fix controller user with swagger

+++ back-end/src/users/users.controller.ts
@@ -14,9 +14,9 @@ import { CreateUserDto } from './dto/create-user.dto';
 import { User } from './entity/users.entity';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { AuthGuard } from 'src/auth/auth.guard';
-import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiTags, ApiResponse } from '@nestjs/swagger';
 
-ApiTags('apis/users');
+@ApiTags('Users')
 @Controller('apis/users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
@@ -24,16 +24,25 @@ export class UsersController {
   @ApiBearerAuth()
   @UseGuards(AuthGuard)
   @Post()
+  @ApiResponse({
+    status: 201,
+    description: 'User created successfully',
+    type: User,
+  })
+  @ApiResponse({ status: 400, description: 'Bad Request' })
   async create(@Body() createUserDto: CreateUserDto): Promise<User> {
     return this.usersService.create(createUserDto);
   }
 
   @Get()
+  @ApiResponse({ status: 200, description: 'Get all users', type: [User] })
   async findAll(): Promise<User[]> {
     return this.usersService.findAll();
   }
 
   @Get(':id')
+  @ApiResponse({ status: 200, description: 'Get user by ID', type: User })
+  @ApiResponse({ status: 404, description: 'User not found' })
   async findOne(@Param('id') id: number): Promise<User> {
     const user = await this.usersService.findOneByid(id);
     if (!user) {
@@ -43,15 +52,31 @@ export class UsersController {
   }
 
   @Patch(':id')
+  @ApiResponse({
+    status: 200,
+    description: 'User updated successfully',
+    type: User,
+  })
+  @ApiResponse({ status: 404, description: 'User not found' })
   async update(
     @Param('id') id: number,
     @Body() updateUserDto: UpdateUserDto,
   ): Promise<User> {
-    return this.usersService.update(id, updateUserDto);
+    const user = await this.usersService.update(id, updateUserDto);
+    if (!user) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    return user;
   }
 
   @Delete(':id')
+  @ApiResponse({ status: 204, description: 'User removed successfully' })
+  @ApiResponse({ status: 404, description: 'User not found' })
   async remove(@Param('id') id: number): Promise<void> {
-    return this.usersService.remove(id);
+    const userExists = await this.usersService.findOneByid(id);
+    if (!userExists) {
+      throw new NotFoundException(`User with ID ${id} not found`);
+    }
+    await this.usersService.remove(id);
   }
 }
